Improve user interface by named actions
Replace '-edxl' options by an explicit command. Updated help texts to
reflect this change and the implementation as well.

A subparser was used to enable cleaner separation of options versus
commands. Options specific to enable / disable / remove are moved into
those options. This means that `ngx-conf -h` shows some a subset of the
available options and that `ngx-conf enable -h` shows the remainder
('-f' and '-r' options were moved due to this).

Using `ngx-conf list` with a file argument will also result in an error
now.

While at it clarify the meaning of "conflicts" for enable and disable.